requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==4.2.4
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
colorama==0.4.6
colorlog==6.7.0
frozenlist==1.3.3
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.48.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==3.4
multidict==6.0.4
openai==0.27.8
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyYAML==6.0
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlmap==1.7.6
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2
